1. add truncate tool that truncates files to the specified byte range. 2. enable explicit command line setting of the number of split and merge workers in the collocation counter. 3. prioritize collocation merges by smallest chunk size. 4. add ngram-freq-filter command line switch that prevents it from reading and outputting totals. 5. factor out count merging out of the ngram counter and make it a generic utility.

+++ src/common/merge.cpp
@@ -62,7 +62,7 @@ void inline printCount(FILE* out, const char* key, const long count)
 
 /// Merges two sorted files with counts into a third sorted file with
 /// the sum of counts. Returns the sum of counts.
-int mergeCounts(FILE* src1, FILE* src2, FILE* out)
+size_t mergeCounts(FILE* src1, FILE* src2, FILE* out)
 { 
   bool hasMore1, hasMore2;
   char key1[buf_size];
